PROJQUAY-39 modified quay liveness and readiness probes
The Liveness and Readniess probes were not working because the connection to port 8443 was being rejected due to SSL cert issues. As the cert is setup for external domain name, requests to `localhost` were being rejected with `400` error code.

The modified check uses `curl` with `-k` (insecure) flag to check if the application is ready and later check if it is still live.

https://issues.redhat.com/browse/PROJQUAY-39

Signed-off-by: Tejas Parikh <tparikh@redhat.com>