Fix syntax error in Python 3.6+
Previously, a library (`trollius`) was used to take advantage of asyncio
in Python 2.7. However, one function (`async`) is now a keyword. This
causes certain parses to encounter a syntax error when parsing the Python
file for linting and other checking purposes.

According to the Python documentation, the old `async()` function is
simply an alias for `ensure_future()` and it has been replaced accordingly.