Catch IntegrityError's in a better way and log
Currently, if we see an IntegrityError, we assume its because the manifest already exists. However, there appears to be a corner case where this is not the case. This change allows us to catch that additional exception, log the original issue, and return a better error to the user

Fixes https://issues.redhat.com/browse/PROJQUAY-219