Catch infinite loops in GC purge code and raise an exception instead
If we have a bug that causes the GC's assumptions to be wrong (such
as an Image added after a purge began), then the purge loop can enter a
state where it will never halt. This change detects that condition and
raises an exception so we can properly investigate the bug.

Fixes https://issues.redhat.com/browse/PROJQUAY-452