Implement a solution for long index names in sqlalchemybridge (#283)
We now cap the index name length at 64 (as per MySQL restrictions). If
an index name is *longer* than that, we hash the index name in its
entirety and then append the short-sha of the hash as a suffix to a
shortened version of the index name. This ensures the index name meets
the length requirements while *also* being stable across generations.

This change also adds a legacy map of long index names to the names we
manually applied, to ensure they don't get regenerated either