chore: remove uploading filtering from imagestorage queries (PROJQUAY… (#768)
* chore: remove uploading filtering from imagestorage queries (PROJQUAY-1914) (#764)

The "uploading" column is an artifact from depending on writing to the
Image table (see BlobUpload table instead). As of 3.4, Quay no longer
writes to that table, and is only needed until quayio moves away from
Clair v2, after which work to remove "glue" code and fully deprecate
the Image table (amongst other tables) can start.

This is done as a separate commit from the actual migration so that it
can be cherrypicked.

* Re-add Image row creation from 4ad5a458

Cherrypicking 4ad5a458 removed writes to the Image tables. For quayio,
this is still needed.