api: prevent unverified users from starting builds
Requires a user to be verified in order to be able to start build of a
repository.

TODO: By default, organization aren't verified. This means that as
part of this change, in addition to newly created orgs, existing orgs
would need to be verified. Which means sending new confirmation emails
for each.