tlscomponent: changing option message for custom tls (PROJQUAY-2428) (#117)
To better reflect the option this commit changes the option presented to
the user. Instead of informing users that Quay pod would be responsible
for terminating the TLS we now inform users that they can provide custom
TLS certificates.

We still do the TLS termination at the load balancer level but this time
using user provided certificates.