notification: check certs exists for webhooks (PROJQUAY-2424)
Make sure the Quay key/cert pair (ssl.key + ssl.cert) have been
mounted/exists whenever it tries sending a webhook request with client
cert. If that's not the case, then the webhook is sent without client
cert. This can happen when TLS is not handled by the Quay container,
in which case these certs are not required by default in the Quay
container.