notification: check certs exists for webhooks (PROJQUAY-2424) (#886) (#900)
* notification: check certs exists for webhooks (PROJQUAY-2424) (#886)

Make sure the Quay key/cert pair (ssl.key + ssl.cert) have been
mounted/exists whenever it tries sending a webhook request with client
cert. If that's not the case, then the webhook is sent without client
cert. This can happen when TLS is not handled by the Quay container,
in which case these certs are not required by default in the Quay
container.

* ci: Update funcparserlib version (PROJQUAY-2520) (#893)

Updating funcparserlib to 1.0.0a0 fixes the issue initially addressed
in d3809b2e.

* ci: pin oci test to a specific sha (PROJQUAY-2570)

This pins the distribution-spec repo to a specific sha
to prevent failing CI because of a change in the source test

Signed-off-by: harishsurf <hgovinda@redhat.com>

Co-authored-by: Harish Govindarajulu <hgovinda@redhat.com>