nginx: support client ip through x-forwarded-for header (PROJQUAY-2883) (#1003) (#1005)
Support client ip from x-forwarded header. This is needed if one of
the intermediate proxy between the client and Quay does not support
proxy protocol.

If set, FEATURE_PROXY_PROTOCOL will use the L4 feature. If not, say,
when a L7 proxy is used, fallback to using x-forwarded-for, and allow
specifying trusted ip/cidr for Nginx.

Co-authored-by: Kenny Lee Sin Cheong <2530351+kleesc@users.noreply.github.com>