healthcheck: add individual endpoints for global services (PROJQUAY-2870)
The status (or endtoend) health endpoint sets its status code to 503
if any of the services being checked is unavailable.

This change adds individual endpoint for each services. This allows a
service like Pingdom to ignore something less critical, like Redis.

Also adds cache status to the global health check. If an external
cache is used, such as redis, it will return its status. If the
default values are used, then the memcached status of whatever node was
hit will be returned.