endpoints/v2: move repository not found check from decorator to endpoint
our api consistently returns 401 when anonymous users cannot access
a resource either due to it being private or when the resource doesn't
exist.

this used to be handled by the `require_repo_read` decorator, but now
with the introduction of pull through proxy organizations (where
pulls to repostories that don't yet exist in our db are allowed) we
had to move the check for repositories not existing to the endpoint code.

IMPORTANT: since the previous behaviour was inconsistent between
anonymous and non-anonymous users, there is no way to fix the
inconsistency without changing API the behaviour.
the changes here will mean that, for anonymous users, instead of
getting 401 when trying to access a repository that doesn't exist they
will get a 404 (which is the behaviour for non-anonymous users).

relates to PROJQUAY-3030