chore: Simplify base image (PROJQUAY-4837)
Downstream build system does not support `FROM base`. To reduce our
burden of maintaining build configurations, we need to eliminate the
base image or at least make it as simple as possible.

The step that builds QUAYDIR doesn't need to use the base image, so it
is migrated to ubi-minimal.