[redhat-3.7] superuser: lowering page limit (PROJQUAY-5178) (#1912)
Lowering the page limit allows for more operations to be done per requested user. Specifically in the scenario where for each user a LDAP lookup is done to check whether the user in the list is a superuser or not. This allows time for the request to complete and load the user list without error.