BREAKING CHANGE: LDAP lookup of robot accounts in the UI for granting                  permission has been dropped. This impacts Users from LDAP                  if they have not logged in already (pre provisioning).
the functions:

* _ldap_single_user_search
* _ldap_user_search

have been updated to provide `context` for dropping the LDAP query. As
the functions are used during login as well as during setting
permissions it is mandatory to be able to distinguish the context.

the functions:

* get_nonrobot_user
* find_user_by_email

from the module `data.model.user` have been utilized to identify, if the
user should be looked up in LDAP or not.