chore: fix build of PyYAML on linux/ppc64le
PyYAML 5.4.1 is the latest (at the moment) version that can be installed
with Cython 3.0 (3.0b9).

As gevent requires Cython 3.0+, we cannot use PyYAML 6.0 that depends on
Cython 0.x (<3.0).

As PyYAML 5.4.1 doesn't restrict the version of Cython that can be used
to build it, pip selects Cython 3.0 when it creates an isolated build
environment and fails to build PyYAML. To fix that, we preinstall build
dependencies and disable the build isolation for PyYAML.