script: Adding python script to automate removal of config editor (PROJQUAY-6707)
The removal of the config editor component from Quay version 3.10 onwards requires manual intervention by users. Customers are obligated to take additional steps. To streamline this process, consider adding the above script to the upgrade process that automatically removes associated objects(such as pod deployments, secrets, routes, and services) of config editor. This script works when quay-operator is deployed in single namespace as well as in all namespaces.