security: change timeout for vulnerability requests (PROJQUAY-7751)
Changes the timeout value for non-indexing requests to clair from 600 to
180.

Because the timeout for a vulnerability report request is so high, heavy
traffic to the security enpdoint results in database connections being
exhausted. Lowering the timeout value should allow requests to complete
and connections to the database to close.