app: Alert user if TESTING is not properly set (PROJQUAY-8123) (#3343)
The `TESTING` flag is used to set Quay in testing mode. In certain occasions (when Quay is deployed on VMs but config tool hasn't verified the validity of the configuration or when users build their own confuguration), this flag sometimes gets omitted from the end `config.yaml` file. When this happens, Quay will start with `TESTING: true` which will disable certain functionality (such as e-mail sending).
We add the check to `app.py` to notify the user that `TESTING: false` is missing or misconfigured in their `config.yaml` file. The change doesn't break Quay startup, merely prints a warning in the container logs when `app.py` gets initialized.