[redhat-3.15] mirror: Add job timeout to mirror configurations (PROJQUAY-7249) (#4051)
Previous global job timeout of 5 minutes was inadequate for big images. The timeout should now be configurable in much the same way as sync is. Minimum job length is 300 seconds (5 minutes), maximum job length is 43200 seconds (12 hours). The option to change skopeo timeout is added to the repo mirroring settings. Additionally, repository mirroring will now correctly react to pressing of the Cancel button in the UI or invoking the cancel mirror API request. When invoked, the sync will be immediately terminated once the current tag finishes upload and will not be queued again until again explicitly enqueued. Previously, pressing the Cancel button would not stop the sync and the worker would still continue to work as if nothing happened.

Push of very large images with very large layers (> 50 GiB) is currently impossible with Ceph/RADOS driver. S3 protocol allows up to 10 thousand chunks to be uploaded per file, with 5 MiB chunk size the maximum theoretical limit for the layer size is 50 GiB. This PR adds an additional optional argument to RadosGWStorage and RHOCSStorage drivers called minimum_chunk_size_mb that allows user to define the upload chunk size. If the parameter is not defined, the driver defaults to 5 MiB.

---------

Co-authored-by: Ivan Bazulic <ibazulic@redhat.com>
Co-authored-by: Ivan Bazulic <ibazulic@laptop.skynet>