alembic: Install certs during alembic migrations (PROJQUAY-9135) (#4125)
Currently, we don't install certs when starting alembic migrations on the db. This poses a problem when Quay is behind a proxy and proxy HTTPS connection is needed to connect to the database. This small fix will make alembic migrations take into account certificates installed in `extra_ca_certs` folder.